+++ webrtc/src/webrtc_c/modules/video_coding/codecs/interface/common_constants.h
@@ -1,0 +1,22 @@
+#pragma once
+
+#include <stdint.h>
+
+#include "../../../../common.h"
+
+#if defined(__cplusplus)
+extern "C" {
+#endif
+
+// -------------------------
+// webrtc::common_constants
+// -------------------------
+
+WEBRTC_EXPORT extern const int16_t webrtc_kNoPictureId;
+WEBRTC_EXPORT extern const int16_t webrtc_kNoTl0PicIdx;
+WEBRTC_EXPORT extern const uint8_t webrtc_kNoTemporalIdx;
+WEBRTC_EXPORT extern const int webrtc_kNoKeyIdx;
+
+#if defined(__cplusplus)
+}
+#endif